Getting ready to implement the auto feature for the morphology reconstruction.

+++ ultraliser/data/morphologies/MorphologyStatistics.h
@@ -54,7 +54,7 @@ public:
      * @brief computeNumberSamplesPerSectionDistribution
      * @return
      */
-    std::vector< float > computeNumberSamplesPerSectionDistribution() const;
+    std::vector< size_t > computeNumberSamplesPerSectionDistribution() const;
 
     /**
      * @brief computeSegmentsLengthDistribution
